Improved Console and fixed .display(..., double batt) bug
The console has many extra functions and an improved help screen. It
also accepts names as well as numbers for input.
Example: /periode 3maanden
/scherm binnenluchtvochtigheid

It searches for keywords, so exact naming is not necessary.

+++ GraadDagen.java
@@ -19,7 +19,7 @@ public class GraadDagen extends Grootheid{
         graadDagen = StatisticsCalculator.graadDagen(list);
     }
     
-    public void display(String periode, boolean knop1, boolean knop2, short batt){
+    public void display(String periode, boolean knop1, boolean knop2, double batt){
         GUIboard.writePageToMatrix("Aantal Graaddagen", graadDagen + "", periode, knop1, knop2, batt);
     }
